+++ cli/src/commands/claim.ts
@@ -1,5 +1,6 @@
 import Command from '../utils/command';
 import terminalLink from 'terminal-link';
+import { getPrimaryDomain } from '../utils/helpers';
 
 export default class Claim extends Command {
   public static description = 'claim an application';
@@ -22,7 +23,7 @@ export default class Claim extends Command {
     const { args } = this.parse(Claim);
     await this.authenticate();
     const app = await this.lycanClient.claimApp(args.token);
-    const domain = app.environments[0].domains[0].name;
+    const domain = getPrimaryDomain(app.environments[0]);
     const link = terminalLink(domain, `https://${domain}`, { fallback: (_, url) => url });
     this.log(`Project successfully claimed! Your project is now available at: ${link}`);
   }